x86: upgrade: integrate efi dualboot support in sysupgrade
Leveraging PARTLABEL support in the kernel, now the boot partition,
which holds both kernels, vmlinuz & vmlinuz_alt, is 'openwrt_boot'.
The main and alternate rootfs partitions are named 'openwrt_rootfs'
and 'openwrt_rootfs_alt'.

If any of these partitions is not found or is found more than once,
sysupgrade errors out explaining you need to have only one of each
on a running system.

The grub bootloader itself is only upgraded when it has changed.

The running system is then checked and both the kernel and rootfs
partition from the upgrade image are copied to the alternate slots
(wih respect to the running ones).

Recommended to run in interactive and verbose mode:

sysupgrade -i -n -p -v $_image

This way it will ask for confirmation before upgrading grub and
before rebooting, which is now optional, ie, we can prepare the
system so the next reboot uses the upgraded firmware but delay
it. We can even disable the upgrade by inverting the default &
fallback entries in /boot/grub/grub.cfg

Signed-off-by: Javier Marcet <javier@marcet.info>