x86: upgrade: add support for pure efi systems & dualboot to sysupgrade.
Leveraging PARTLABEL with openwrt_boot holding the boot partition,
openwrt_rootfs & openwrt_rootfs_alt the main and alternate rootfs parts.
The main and alternate kernels are named vmlinuz & vmlinuz_alt.

If the any of the partitions is not found or is found more than once,
the script errors out explaining you need to have only one of each on
a running system.

The grub bootloader itself is only upgraded when it has changed.

The running system is then checked and both the kernel and rootfs
partition from the upgrade image are copied to the alternate ones.

Recommended to run in interactive and verbose mode:

sysupgrade -i -n -p -v $_image

This way it will ask for confirmation before upgrading grub and before
rebooting, which is now optional, ie, we can upgrade a running system,
but reboot at a later time, maybe with a new upgrade. Disabling the upgrade
when the reboot is not done is as easy as inverting the default & fallback
entries in /boot/grub/grub.cfg

Signed-off-by: Javier Marcet <javier@marcet.info>