x86: drop hybrid BIOS mode from EFI images in preparation for dualboot
With this in place a native EFI system will be able to sysupgrade
a running system, plus it will keep the previous image which is easy
to switch to in case of a problematic upgrade.

We now leverage PARTLABEL thus it is not configurable anymore.

We change the GUID types of the partitions:

 - Drop the BIOS_BOOT partition, it is only needed for hybrid mode
 - Add new GUID, FREEDESKTOP_BOOT 0xea, for the boot/kernel partition
 - The main and alternate root fs, labelled "openwrt_rootfs" and
   "openwrt_rootfs_alt", get GUID_PARTITION_LINUX_FILESYSTEM.

Signed-off-by: Javier Marcet <javier@marcet.info>