x86: drop hybrid BIOS mode from EFI images in preparation for dualboot
With this in place a native EFI system will be able to sysupgrade
a running system, plus it will keep the previous image which is easy
to switch to in case of a problematic upgrade.

We now leverage PARTLABEL, as such it is not configurable anymore.

We change the GUID types of the partitions: we drop the BIOS_BOOT
partition, since it is not needed without hybrid mode; we add a new
FREEDESKTOP_BOOT GUID and we replace GUID_PARTITION_BASIC_DATA with
GUID_PARTITION_LINUX_FILESYSTEM.

The boot partition, with PARTLABEL=openwrt_boot, holds the kernels
(vmlinuz and vmlinuz_alt). The main and alternate rootfs partitions
have PARTLABEL=openwrt_rootfs and PARTLABEL=openwrt_rootfs_alt.

Signed-off-by: Javier Marcet <javier@marcet.info>