x86: drop hybrid BIOS mode from EFI images, decoupling completely grub2 and grub2-efi images in preparation for dualboot.
With this in place a native EFI system will be able to sysupgrade
a running system, plus it will keep the previous image which is easy to
switch to in case of a problematic upgrade.

We now leverage PARTLABEL, as such it is not configurable anymore.

We change the GUID types of the partitions. We drop the BIOS_BOOT
partition, add a new FREEDESKTOP_BOOT GUID with PARTLABEL=openwrt_boot
to hold the kernels (vmlinuz and vmlinuz_alt), change GUID_PARTITION_BASIC_DATA
to GUID_PARTITION_LINUX_FILESYSTEM with PARTLABEL=openwrt_rootfs and
PARTLABEL=openwrt_rootfs_alt for the main and alternate rootfs.

Signed-off-by: Javier Marcet <javier@marcet.info>