base-files: improve /var as permanent directory
Users may want a permanent /var instead of a symlink to /tmp.
In this situation, some dirs in /var should still be stored in /tmp,
as they store files that are not intended to survive reboot.

This commit creates symlinks for /var/etc, /var/lock, /var/run and /var/state
and their targets in /tmp, when /var is normal directory.

Signed-off-by: Marius Dinu <m95d+git@psihoexpert.ro>